Added login screens (3 xml or all in one) AND Followers search setup
Dropdown search for a UserName to follow, shows matching names on server... Then: The login screen has a new user text input (Login by clicking avatar?), a textview for stating the username already exists, and a textview for welcoming the user if they've already logged in before. May be able to make them invisable in activity, if not just change the view in the activity and remove the extra buttons.